src/bootchooser: split backends into dedicated files
This splits the five backends into dedicated files (barebox, custom,
efi, grub, and uboot).

This change implies:
 - to create source and header files for every backends
 - to include backend headers to bootchooser.c and use the prefixed
   version of backend functions
 - to share function bootchooser_order_primary()
 - to adds new sources to meson.build
